Track PSBT consistency check in type
This change tracks whether PSBT was verified to be consistent (having
correct counts of inputs and outputs) in type. It moves panic closer to
PSBT construction and it found a possible error - original PSBT is not
guaranteed by the caller to be consistent but we relied on it without
documenting it.

This is an MVP to close #5