Add sender's additional_fee, not an estimate
Rust bitcoin does not yet give us tools to estimate A PSBT's
transaction fees out of the box. Inordertodo  that, it would need to
estimate in dummy witness / signature data to contribute to the final
weight. Adding the additional fee and ignoring any fee contribution of
our own for a first pass eliminates the need for an estimate.

Adds weight, feerate from rust-bitcoin.