Add amount checks to integration tests
At the end of each integration test, check the resulting transaction
inputs/outputs, and the resulting balances for the sender and receiver.

Note that the sender sweep test case requires the original PSBT fee rate
to be high enough to cover the receiver input, because neither the
sender nor the receiver can contribute additional fees in the payjoin
PSBT due to current API limitations.