Remove MaybeMixedInputScripts typestate
The BIP78 spec says:
"If the sender's inputs are all from the same scriptPubKey type, the
receiver must match the same type. If the receiver can't match the type,
they must return error unavailable."

Instead of rejecting an original PSBT that contains mixed input scripts
types, we should be checking if all input script types are the same and
only error if the receiver input contributions would *introduce* mixed
script types.

This check is now done in `contribute_inputs` (for v1 receivers only)
and the previous mixed input scripts check is removed.