replace bip21 crate with bitcoin_uri
The `pj` parameter of the BIP 21 URL is itself a URL which contains a
fragment.

The # character was not escaped by bip21 during serialization. According
to RFC 3986, this terminates the query string, effectively truncating the `pj`
parameter.

The buggy deserialization likewise ignored #, parsing it as part of the
value, which survived a round trips with the incorrect serialization
behavior.