Encode fragment parameters as bech32
These changes are to facilitate more efficient QR encoding, by
exploiting the alphanumeric encoding mode.

Use `+` instead of `&` as a parameter delimiter, as only the former is
in the QR alphanumeric character set.

Instead of encoding keys and values using `=` as a delimiter (which also
not in alphanumeric set), the key is the human readable part of a bech32
string with no checksum. This effectively makes the `1` a key/value
delimiter.

The expiry time is serialized as a bitcoin consensus u32 value (little
endian), and also encoded as a bech32 string instead of a decimal value.