Have directory return bound port on listen
ensure that the db: testcontainer::Container variable does not go out of
scope while the directory is running.

previously the directory task itself was awaited on by init_directory,
whereas in the modified code it is instead returned as part of the result
due to the different return value of listen_tcp_with_tls_on_free_port.
this indirection de-coupled the db variable's lifetime from that of the
directory, allowing it to go out of scope earlier than expected.

Co-authored-by: Yuval Kogman <nothingmuch@woobling.org>
Co-authored-by: spacebear <git@spacebear.dev>