Encode subdirectory ID as bech32
No checksum or human readable part is included, resulting in 13
uppercase characters, which is amenable to QR code alphanumeric mode.

In the directory, Instead of parsing subdirectory IDs to obtain a byte
array, simply treat them as strings again. This is simpler than
introducing a dependency on the payjoin workspace crate in order to
parse them back into the ShortId type, since this is likely to be a
negligible optimization of the redis serialized key size, and does not
affect the key space.