Fix unresumable Sender
The Sender was generating new hpke keys on resume since keygen was being
done in `extract_v2`. This caused a problem where when the Sender was
persisted, stopped, and resumed, the receiver would have already pushed a
response to a different ShortId than the one the Sender would look for it
in the resumed state.

By generating a key on Sender creation and persisting that the Sender can
produce a consistent HpkeContext every single run.