Replace PjUriBuilder with mutable PjUri
The builder pattern was feature gated and confusing when there were
really only two distinct ways to build Uris:

1. From plain parameters in v1, from which to add amount, label, etc.
2. From a v2 receiver, from which to add amount, label, etc.

amount, label, message, etc. can be set by mutating the bitcoin_uri::Uri
struct directly, and we weren't enforcing anything special like pjos
in the builder that we could not in the v2::Receiver directly. So we can
delete code and have a simple interface that gets us closer to additive
v1/v2 features.