Remove ohttp_relay from SessionContext
The context may use multiple OHTTP Relays, so pass it in the extract
method instead.

Our integration tests currently fail to relay messages through the OHTTP
Relay to the Directory, so they need to pass messages directly to the
directory. The reason they're failing otherwise is because the OHTTP
Relay only forwards requests using rustls with_webpki_roots, the
tests' directory Urls use HTTPS, and the webpki_roots do not contain
or allow self-signed certificates. I.e. we have not equivalent
_danger-local-https feature available in ohttp_relay and the request
isn't downgraded to plain HTTP, so it fails to relay.

We may consider having extract_req functions downgrade OHTTP requests
to directory targets to HTTP, because they are protected with e2ee as
part of OHTTP outside of HTTPS.

This change continues to use mock_ohttp_relay (directory) Urls until
such a change is made.