Use FeeRate::BROADCAST_MIN for minfeerate defaults
Relay policy's BROADCAST_MIN is far more usual than FeeRate::MIN (ZERO)
when it comes to thinking of minnimum feerates to paramaterize.

Use the default in payjoin-cli since it is a reference implementation.

This is only a default. If some implementation wants to use a lower fee
rate they still may.

Close #489