Introduce IntoUrl for ergonomics
Accept stringly types in  typestate methods so that downstream
implementations may depend on the typestate machines for URL parsing.

This reduces a step for them and unifies our error handling for added
internal complexity, which seems like the express purpose of the library.

Create new ParseUrl error types in our state machines.