Make payjoin-cli features additive
Make use of the AppTrait and a `--bip78` cli argument to allow both to
be compiled into the same binary and a user to choose the protocol at
runtime.

Parse only one config version "mode" at runtime so that errors in a
version unused at runtime will not cause errors.

Use a determine_version fn in config so that the default behavior
without an explicit version flag runs the highest version available.
Use argument matches in main to reduce presedence if multiple versions
are available.