Create new lowercase error catch for uri endpoint fragments
The pj= param in the pjuri contains multiple different parts but there exist
some gotchas when trying to use them as the url path can be parsed
even when lowercase but the fragment object cannot.
This means that it is possible to copy or display a pjuri that will
not work if it is simply converted to lowercase.
This pr does not create any method to force the fragment to be uppercase
as I think it really should be on the implementation to make sure that
casing is correct, so I just created the case error catch.