Depend on bitcoin/rand-std for thread_rng in v1
`cargo check --no-default-features --features v1` was producing an error
because `thread_rng` was not available with the given feature set.
Somehow the other feature combinations make std available.

no-std randomness would be nice for web browser embed, but we should
cross that bridge when we get to it by using the `getrandom` crate with
`js` feature enabled. For now all of our targets can depend on `std`.