Fix confusing and incorrect test condition
When a v2 receiver polls expecting a 200 or 202 status response, that
status code is associated with the OHTTP encapsulated response, and is
therefore handled in the process_res method and detected implicitly via
the empty body branch.

The response status for OHTTP gateway requests will always be 200
regardless of what the internal request's status is.