Add `enum ErrorCodes`
Well-known error codes are used on both sides of the protocol. Having
a strong type will let us serialize an error Reply across a foreign
language boundary.

We were returning a non-spec sender-params-error which amounts to a
rejection of the original PSBT for unparseable feerate. Instead,
I decided to return the well-known original-psbt-rejected error code
and to include the parse error in the debug message in order to comply
more closely with existing spec.