Fix bdk_integration_test
This replaces the `payjoin-directory` dependency with
`payjoin-test-utils` which allows us to use the test fixtures instead of
running a live server and duplicating a bunch of code.

Parts of the test are corrected (e.g. setting content-type headers in
all requests, which was done inconsistently previously and caused 4xx
errors) and it is now more inline with the v2_to_v2 integration test
from the `payjoin` library.