Create fallback ohttp-relay logic for payjoin-cli
The reliance on ohttp-relays can cause some distruption
if they ever go offline but we did not have any examples of
a fallback mechanism for choosing the relays at random and falling back
to other relays in case they fail.

This adds a new RelayState to manage which relays have already been tried
and ulitmately hangs on to the first successful relay that it reaches.