Persist directory key configuration
Include functions for generation and reading from persistence. Note KeyId = 1
Is still used and there is no function for key rotation yet.

Persist a key on first run if none already exists.

The OHTTP Server configuration is now an argument to listen functions.
For now, a default path of `ohttp_keys/1.ikm` is used to persist the key
material. "1" is the KeyId. The default path is relative to where the
binary is run.