Create a duplicate proposal check in ns1r
When building a multiparty proposal we did not have a catch to
make sure that the proposals were all unique and as such a multiparty
could be built by just adding the same proposal over and over.  This creates
a new error IdenticalProposals that checks to make sure the contexts are not
matching when adding to a multiparty.