Merge parallel feature flag local certificate logic in cli ohttp
These ohttp functions had parallel methods when running tests vs running
in production. This made it difficult to read as there were two
functions that appeared to do completely different things when a feature
flag was present.

Because localhost relays are not actually able to be proxies here we
still need to opportunistically select them from the config when running
our tests.