Lock cargo mutants version to 25.0.1
The new version of cargo mutants introduced a set of new mutants that
ultimately is not worth the added effort to fix at the moment as well as
introducing some breaking changes that have not been clarified yet how
to work around them. For now I am locking the cargo mutants version
until we have more clarity on the stability of the mutants workflow and
we can put more priority into full mutation coverage.