Update FFI transition objects to return PersistedError
`PersistedError` encapsulates the api side errors as well as storage
related errors that can occur during a state transition. The following
commit updates all the sender side transitions objects to return
`PersistedError`. There was a naming conflict with the reciever type.
Both error types had to be renamed and prefixed with
"Sender|Receiver_...".