Capture hpke reply key in session event
The session reply key stays `None` until we parse it from an
`UncheckedProposal`. We lose this information because we don’t store an
updated `SessionContext` in the `UncheckedProposal` session event and we
don’t want to store the full updated context either. Instead, we stick
to storing only what the session learns at that point. In this case,
that’s the proposal itself and the optional HPKE reply key.