Deduplicate multiparty response error with ohttps DirectoryResponse
In #660 we added a process_get_res and process_post_res methods in
ohttps to pull away some of the logic from our send module but
multiparty remained the same.

This PR attempts pull sync the multiparty sender error handling with
this new approach. As well this change gets rid of the remaining mutant
on this status code match block.