Add CONTRIBUTING.md
Adds a `CONTRIBUTING.md` file to the `.github` directory, adding
guidelines for creating issues, writing documentation, and contributing
code as discussed in #758.

The decision to place it in `.github` was made due to the fact that it's
already being used for another [community health
file](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file)
([`ISSUE_TEMPLATE`](https://github.com/payjoin/rust-payjoin/tree/master/.github/ISSUE_TEMPLATE)).

This also moves the "Contributing" section of the `README` to
`CONTRIBUTING`, and adds a link to it in its' place.