Create new test to catch non-deterministic check_outputs mutation
In the send module, check_outputs had a mutation whereby the match arm
was skipped and the logic in our tests to prevent wallet fingerprinting
was causing the error to not be caught consistently.

By creating a test with fixed psbts we can esentially force the mutation
to be caught.