Create new test to catch non-deterministic check_outputs mutation
In the send module, check_outputs had a mutation whereby the match arm
was skipped and the logic in our tests to prevent wallet fingerprinting
was causing the error to not be caught consistently.

By createing a test where the Sender output is placed first we can
esentially skip this mutated logic forcing the mutation to be caught.