Expose fallback tx off receiver session history
If a `MaybeInputsOwned` session event is present in the log then we know
the session has persisted a "broadcastable" fallback tx by virute of
successfully transitioning past `UncheckedProposal` via calling
`check_broadcast_suitability` and persisting the result. This commit
explicitly does not use the unchecked session event because the fallback
at that typestate may fail consensus or policy checks.