Patch ffi payjoin-directory dependency
This patch changes the dependency from using a published version from
crates.io to using a local path dependency. This ensures that we can
still compile the ffi crate with newer versions of the test-utils (which
depends on payjoin-directory) even if the newer versions of the
workspace dependencies are not available on crates.io.