Use `-` as fragment delimiter instead of `+`
Although RFC 3986 (URIs) does not assign any special meaning to `+`, in
fragment parameters or in general, RFC 1866 (HTML 2.0) section 7.5 uses
it as a delimiter for keywords in query parameters. As a result some URI
libraries interpret `+` in URIs as ` `, even in fragment parameters.

Although not insurmountable (such transformation BIP 77 URIs is
reversible because ` ` is not used and `+` was only used for fragment
parameter delimitation) this presents friction and is in general
confusion, so to improve compatibility with such libraries `-` is now
used instead. It has no reserved meaning as a sub-delimiter.

For the time being when parsing both `+` and `-` will be accepted, but
only `-` will be used when encoding fragment parameters.