Add python devShell to nix flake
Adds a top-level python devShell to the nix flake which creates a python
virtual environment with all necessary dependencies installed.

Now, assuming Docker is running and nix is installed on the system, a
developer can simply run `nix develop .#python` and have a shell with
all the requirements to run `python -m unittest --verbose` setup,
or import the `payjoin` package (locally built) to their project.