Check fees in multiparty process_proposal
The distinction beetween 1s1r async payjoin and ns1r is that the sender
checks the proposal before the receiver signs in the first round
of communication, so the receiver inputs won't yet be finalized. Therefore,
in order to share functionality, the receiver input finalized check needs
to be triggered by a condition. This does that.