Use FnMut closures for v1 receiver typestate checks
In our liana wallet integration we are encountering some issues due to
how the liana devs have setup their database, namely that their db
always references a mutable self even on read calls within the db so we
need to have some mutability when accessing things like an address or
outpoint from the db. This was not possible when our closures were
limited to non-mutable.

I considered using FnOnce but we have some map methods that prevent us
from doing that as map have the possiblity of forcing the closure being
called more than once.