payjoin-cli: handle binding free port in v1 server
When binding a socket address with port 0, the OS will allocate a free
port. The --port CLI argument supports doing this, but the bound port
number was not used in any way.

This change checks that if --port 0 is specified, --pj-endpoint does not
specify a port, and sets the port to the assigned to the listener by the
OS before generating payjoin payment URI.