payjoin-directory: introduce db::Db trait
This is a refactoring of the existing redis DbPool, simplifying it a
bit and renaming the various methods for clarity.

The trait is async, with its futures also satisfying `Send` in order to
allow its use in the `hyper::Service` trait.

DbPool was renamed to db::redis::Db (reexported as RedisDB) in
anticipation of additional impls of the trait.

Co-authored-by: benalleng <benalleng@gmail.com>