[WIP] payjoin-directory: introduce db::mem
TODO:
- cleanup invariant checks

This database is purely in memory, consisting of a hashmap for stored
payloads (v2 payloads and v1 request payloads), and two waitmaps for
notifying waiting readers of new payloads (and v1 responses).

Mailbox contents are retained for up to a week by default, or up to 24
hours at capacity. Read mailboxes have a grace period of 10 minutes
before they are expired. v1 responses are not stored at all and are
only delivered to a single waiting v1 sender.

No concurrent hashmap is used, because concurrency is implicitly bound
by bitcoin's transaction throughput limits. Because only mutual
exclusion for the entire structure is used, pruning is done on a per
need basis by the locking thread and not in a background task.