Filesystem based mailbox database
v2 payloads are saved to disk.

v1 requests and responses are not, because they only make sense to keep
around so long as the sender's connection is still active, and because
they are not encrypted.

In flight requests (long polls) are tracked waitmaps for notifying
waiting readers of new payloads or v1 responses.

Mailbox contents are retained for up to a week by default, or up to 24
hours at capacity. Read mailboxes have a grace period of 10 minutes
before they are expired.

For simplicity no concurrent hashmap is used, and the mutex is held
during writes to disk. This is because concurrency is implicitly bound
by bitcoin's transaction throughput limits, for the retention times
described above a very generous upper bound is 25 writes per second.
Because only mutual exclusion for the entire structure is used, pruning
is done on a per need basis by the locking thread and not in a
background task.