Correct stale mutant exclusions
There was a rebasing mistake on a previous commit as I created some
exclusions but their name was changed before #910 was merged and
they were not caught.

There is one exclusion that is simply renamed but another I decided to
change the logic such that the exclusions are no longer possible.