Add `Implementation` error to v2::SessionError
In the async. payjoin context, we need a mechanism to represent
integration errors that should not be relayed to the
counterparty. Currently, these errors are incorrectly represented as
`ReplayableError::ImplementationError`.

This commit corrects the error handling by updating `finalize_proposal`
to return `SessionError::IntegrationError`, which provides a more
accurate and semantically correct representation for integration
failures. Moreover, `SessionError`s cannot be accidentally converted and
relayed to the sender as a json reply as they do not impl the
conversion.