Use reply key for replyable errors to v2 senders
Before some relatively recent spec changes, two messages were written to
the same mailbox. The error path was not updated to reflect this, so the
sender will not receive it (as it is polling a different mailbox), and
the receiver further relied on the directory allowing mailbox contents
to be overwritable (or at least accepting the POST request).

With this change a replyable error to a v2 sender will be sent to the
reply key, assuming one was obtained from the first message. If no key
can be recovered from the message then there is no point in sending a
reply as the peer is not adhering to the protocol.