Remove `Uninitialized` as a session variant
Unintialized should be an unreachable state and we are misusing it
today. If one does not have any events in their log and replays then
they will receive an uninitilaized receiver. Which is not useful. Not
having any events and deciding to replay an session is most likely a bug
in the application and should get an error (NoEvents). Additionally, bc
sessionState is pub applications have to match on uninitialized when
processing their receiver states -- instead you should receive an error
when you have no events in your log. The other reason Unintialized
existed is to having a starting state when replaying the SEL. Which we
can resolve by having a special carve out for the first sessionState
(SessionState::new).