Replace SystemTime with bitcoin::absolute::Time
Internally, introduce a newtype wrapping bitcoin::absolute::Time (which
is what BIP 77 specifies) instead of using std::time::SystemTime.

A fallible constructor `Time::from_now(std::time::Duration)` is used to
construct these values from a `Duration`, which is indirectly called by
`with_expiry`. No `IntoTime` trait analogous to `IntoUrl` is included in
this commit, but this could be added there as long as both the
re-exported and the new `Time` type implement it.

Introducing a newtype lets us add additional conversion and arithmetic
functionality later without breaking the API or relying on extension
traits on the bitcoin Time type. Since it is now only `pub(crate)`, the
newtype wrapper's existence is now not entirely justified, as this no
longer re-exports a foreign type in our pub api (previous iterations of
the commit made it pub, replacing the use of `Duration`).

Co-authored-by: Yuval Kogman <nothingmuch@woobling.org>