Save max fee rate in receiver session context
The reference implementation currently overrides the original max fee
rate preference when resuming a session. This commit pins the max
fee rate preference to the session by persisting it in the session
context. When applying fees, if the applications’s fee rate
preference has changed, the updated preference takes precedence.

Co-authored-by: Johnosezele <johnosezele@gmail.com>