Only throw Expired error from protocol operation
Do not throw an Expired error as a ReplayError variant. An expired
session does indicate an error with replay. Rather, a replay of an
expired session that has not yet reached a terminal state will produce
an error when it follows protocol and tries to create a request. That
error will create a SessionEvent::SessionInvalid (or whatever that type
turns into after follow ups) which will later be replayed by the state
machine.

This separates the concerns of the protocol from the session replay and
more accurately reflects the protocol operation. This lets us record to
what extent the protocol actually was able to operate before expiration
since all of the events will be replayed up to the terminal expiration
event.

I noticed that the receiver's final post was not checking expiration, so
I added and tested that as well after removing the replay expiration
test.

Make SessionError(pub(super) InternalSessionError) so that the error
variant can be used in matches in unit tests.