Add receiver monitor typestate
A payjoin session is closed once there is nothing more to do in the
protocol but wait for expiry, or confirmation of one of the possible
transactions to settle the outcome.

This commit introduces a monitoring typestate as the final receiver
typestate. Monitoring checks for the confirmation of session
relevant transactions. Concretely we check for:
* The payjoin proposal via txid if all inputs are segwit.
* The fallback tx via txid.
* The payjoin proposal via spent outpoints if at least one input is not
  segwit.
* Any spent outpoint which would indicate a double spend of an outpoint
  that was used in the payjoin.

The typestate accepts two callbacks:
* `transaction_exists(txid)`
* `outpoint_spent(outpoint)`

If a broadcasted payjoin proposal was detected the session will close
and the success event will include the sender `scriptsig` and/or
`witness`. Fallbacks and double spends will also close the session and
the event logged will include which condition closed the session.