Remove Url from public API
This commit removes the Url from the public Api where it is an arg in
public api methods and also as an impl in the IntoUrl interface so that
it is not accepted where IntoUrl is used.

This commit does not remove Url internally from the payjoin crate or
remove the use of Url in any external crates, it simply transitions the
type to a `&str` or `String` when passing into payjoin methods.