Replace Url dep with payjoin Url for public Api
This adds a payjoin Url with an IntoUrl interface that prevents The Url
dep from being publicly accessible in our Api.

Some notable changes in the payjoin crate are that the Request Struct
now returns a String on the url field instead of a Url directly.

The cli crate uses this Url Struct directly where possible and
parses from a `url::Url` where usage of methods like `port()` are
required.

The ffi crate uses Strings directly and does not use the ffi specific
Url wrapper as Strings already implement lift and lower and do not need
any special wrappers for bindings in this case.  Notably without the Url
ffi struct the unittests that check for a valid Url struct now just
check for an existing `url.query` field.

The test-utls crate remains unchanged and still returns a `url::Url`
where appropriate.