Move bech32 decoding errors to OH param error
Treating this as a param, not value error makes more sense since the HRP
is used in fragment parameter extraction.

Additionally the internal errors are not very useful, because we do not
expect URIs to be transcribed manually. Therefore to simplify the public
API these foreign types are no longer exposed in our error types and
both of these just become `InvalidFormat`.