InvalidFormat instead of DecodeBech32 & InvalidHrp
When parsing URIs it is unlikely that these errors will provide any
benefit to application developers or users, since we do not expect URIs
to be transcribed manually.

InvalidHrp should not happen at all, because fragment parameter
extraction depends on the prefix and separator to begin with, so any
invalid HRP should manifest as a missing parameter.

DecodeBech32 would also normally not happen, we don't use checksums for
these values so transposition or substitutions would not raise an error.

Removing this detail in our errors lets us avoid re-exporting types in
our public APIs.