Panic on RwLock errors in FFI wrappers
Crashing the process on a concurrent access attempt seems acceptable
because it's a concurrency issue on the host's side; only one process
should be progressing the payjoin session state. And mapping this type
of error to ImplementationError suggests that the error is recoverable
when it's actually a fatal programming error.