Fix apply_fee_range error handling
The FIXME comment says this should return a terminal error, but as far
as I can tell the only way that method fails is if the receiver's fee
contributions would exceed their specified `max_effective_fee`. By
making this a transient error we allow a receiver to try applying fee
contributions again with a different feerate.