Simply return replay error on expiration
Adhering to the principle that only primary information becomes an
event, knowing the session's expiration (in SessionContext) is
sufficient and no new information implies no outcome event.

The state machines already adhere by this and only returns an error
without resulting in a state transition. This commit applies the same
logic to `replay_event_log` for sender and receiver.

Even assuming a non-monotonic system clock, the implication of this
change is that a session that was previously considered expired may turn
out to not actually be expired, and is allowed to resume. This is
desirable behavior if e.g. the clock being incorrectly set to a future
time was what led to the session being expired in the first place.