Close session on fatal error only when specified
Only a subset of fatal errors should result in immediate closure of a
receiver session. In many cases an attempt should first be made to
respond to the sender with an error as specified in BIP78.

Instead of scrutinizing various error types to determine whether it
should be replyable or not, the callsite can simply provide the
appropriate SessionEvent::HasReplyableError(JsonReply) for replyable
errors, or SessionEvent::Closed(Failure) for non-replyable ones.