Introduce state transitions to `HasReplyableError`
Only a subset of fatal errors should result in immediate closure of
areceiver session. In many cases an attempt should first be made
torespond to the sender with an error as specified in BIP78.

Instead of scrutinizing various error types to determine whether
itshould be replyable or not, the callsite can simply call the
appropriate state transition method: `replyable_error()` for replyable
errors or `fatal()` for non-replyable ones. For replyable errors, the
resulting HasReplyableError typestate can be obtained from the resulting
`PersistedError` with the `error_state()` method.