Demonstrate failure handling with fallback TX broadcast
As a refrence implementation , payjoin-cli needs to demonstrate
proper fallback transaction handling in the case of payjon failure.
From what i see , both sender and receiver can broadcast the fall
back transaction , This pr starts by handling the sender side
of things and would eventually progress to the receiver .

this addresses #1156

This is mostly sender fallback handling.

Broadcast on Sessionoutcome